ClearTextButtonBehavior to acheive a better handling
