+++ MahApps.Metro/Styles/Controls.PasswordBox.xaml
@@ -1,7 +1,9 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
                     xmlns:Controls="clr-namespace:MahApps.Metro.Controls"
-                    xmlns:Converters="clr-namespace:MahApps.Metro.Converters">
+                    xmlns:Converters="clr-namespace:MahApps.Metro.Converters"
+                    xmlns:i="http://schemas.microsoft.com/expression/2010/interactivity"
+                    xmlns:behaviours="clr-namespace:MahApps.Metro.Behaviours">
 
     <ResourceDictionary.MergedDictionaries>
         <ResourceDictionary Source="pack://application:,,,/MahApps.Metro;component/Styles/Shared.xaml" />
@@ -172,7 +174,11 @@
                                     Content="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=(Controls:TextBoxHelper.ButtonContent), Mode=OneWay}"
                                     Visibility="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=(Controls:TextBoxHelper.ClearTextButton), Mode=OneWay, Converter={StaticResource BooleanToVisibilityConverter}}"
                                     Width="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=(Controls:ControlsHelper.ButtonWidth), Mode=OneWay}"
-                                    IsTabStop="False" />
+                                    IsTabStop="False">
+                                <i:Interaction.Behaviors>
+                                    <behaviours:ClearTextButtonBehavior />
+                                </i:Interaction.Behaviors>
+                            </Button>
                             <ContentPresenter x:Name="PART_CapsLockIndicator"
                                               Grid.Row="0" 
                                               Grid.Column="1"
@@ -426,7 +432,11 @@
                                     FontFamily="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=(Controls:TextBoxHelper.ButtonFontFamily), Mode=OneWay}"
                                     Content="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=(Controls:TextBoxHelper.ButtonContent), Mode=OneWay}"
                                     Template="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=(Controls:TextBoxHelper.ButtonTemplate), Mode=OneWay}"
-                                    IsTabStop="False" />
+                                    IsTabStop="False">
+                                <i:Interaction.Behaviors>
+                                    <behaviours:ClearTextButtonBehavior />
+                                </i:Interaction.Behaviors>
+                            </Button>
                             <ContentPresenter x:Name="PART_CapsLockIndicator"
                                               Grid.Row="0"
                                               Grid.Column="1"
